+++ src/lib/google-cloud-storage/index.ts
@@ -1,4 +1,3 @@
 export { InjectGoogleCloudStorage } from './google-cloud-storage.decorator';
 export { GoogleCloudStorageModule } from './google-cloud-storage.module';
-export { GoogleCloudStorageConfig } from './google-cloud-storage.dto';
 export { GoogleCloudStorage } from './google-cloud-storage';