+++ src/lib/mongoose/index.ts
@@ -1,2 +1,2 @@
-export { InjectMongoose } from './mongoose.decorator';
+export { InjectModel } from './mongoose.decorator';
 export { MongooseModule } from './mongoose.module';