+++ src/lib/typeorm/typeorm.module.ts
@@ -1,6 +1,5 @@
 import { Global, Module } from '@nestjs/common';
 import { TypeOrmModule as OrmModule } from '@nestjs/typeorm';
-import * as _ from 'lodash';
 
 import * as Entities from '@entities';
 import * as Repositories from '@repositories';
@@ -11,7 +10,7 @@ import { TypeOrmConfigService } from './typeorm.service';
 @Module({
   imports: [
     OrmModule.forRootAsync({ useClass: TypeOrmConfigService }),
-    OrmModule.forFeature([..._.map(Entities, x => x), ..._.map(Repositories, x => x)])
+    OrmModule.forFeature([...Object.values(Entities), ...Object.values(Repositories)])
   ],
   exports: [OrmModule]
 })