+++ src/schema/index.ts
@@ -1,3 +1,0 @@
-export * from './audit.schema';
-export * from './client-app.schema';
-export * from './user.schema';